NET-2: moved packet to phase 3 packet file
